chore: rename Node to EmberNode and Fucntion to EmberFunction - missed a couple
