+++ src/encoders/ber/EmberElement.ts
@@ -4,11 +4,11 @@ import { encodeCommand } from './Command'
 import { Command } from '../../model/Command'
 import { encodeParameter } from './Parameter'
 import { Parameter } from '../../model/Parameter'
-import { encodeNode } from './Node'
+import { encodeNode } from './EmberNode'
 import { EmberNode } from '../../model/EmberNode'
 import { encodeMatrix } from './Matrix'
 import { Matrix } from '../../model/Matrix'
-import { encodeFunction } from './Function'
+import { encodeFunction } from './EmberFunction'
 import { EmberFunction } from '../../model/EmberFunction'
 import { encodeTemplate } from './Template'
 import { Template } from '../../model/Template'
